@@ -30,6 +30,8 @@ import java.util.List;
 import java.util.Locale;
 import java.util.regex.Matcher;
 
+import androidx.core.view.MenuItemCompat;
+
 public class DictionaryEntryDetailFragment extends DetailFragment {
 
     private static final String TAG = DictionaryEntryDetailFragment.class
@@ -156,8 +158,7 @@ public class DictionaryEntryDetailFragment extends DetailFragment {
         inflater.inflate(R.menu.dict_detail, menu);
 
         MenuItem menuItem = menu.findItem(R.id.menu_dict_detail_share);
-        shareActionProvider = (ShareActionProvider) menuItem
-                .getActionProvider();
+        shareActionProvider = (ShareActionProvider) MenuItemCompat.getActionProvider(menuItem);
 
         if (shareActionProvider != null) {
             shareActionProvider.setShareIntent(createShareIntent());
