@@ -46,7 +46,7 @@ public class WwwjdicPreferences extends PreferenceActivity implements
     public static final String PREF_AUTO_SELECT_MIRROR_KEY = "pref_auto_select_mirror";
     public static final String PREF_WWWJDIC_MIRROR_URL_KEY = "pref_wwwjdic_mirror_url";
     public static final String OLD_DEFAULT_WWWJDIC_URL = "http://gengo.com/wwwjdic/cgi-data/wwwjdic";
-    public static final String DEFAULT_WWWJDIC_URL = "http://www.edrdg.org/cgi-bin/wwwjdic/wwwjdic";
+    public static final String DEFAULT_WWWJDIC_URL = "https://www.edrdg.org/cgi-bin/wwwjdic/wwwjdic";
 
     public static final String PREF_WWWJDIC_TIMEOUT_KEY = "pref_wwwjdic_timeout";
     private static final int WWWJDIC_TIMEOUT_DEFAULT = 10 * 1000;
@@ -160,6 +160,7 @@ public class WwwjdicPreferences extends PreferenceActivity implements
     protected void onStart() {
         super.onStart();
 
+        // TODO: migrate to PreferenceFragment?
         getActionBar().setDisplayHomeAsUpEnabled(true);
     }
 
